Fixed bug #5
This was a very subtle bug if I may say so myself. The data was all
being changed but the problem is that the view wasn't being notified of
the feedback change because the view does not monitor each individual
guess. Instead it monitors the play list. On the ProvideFeedback command
the actual Guess was being modified instead of the PlayList model
through its addFeedbackToLastGuess() this method actually notifies all
the views of it's inmediate change. While the data was changed on the
back the view wasn't being updated twice like it should: Once for the
new code and another time for the new feedback.